Customer, Seller, User en Cascade

+++ src/main/java/com/erp/inventariapp/Services/UserService.java
@@ -11,6 +11,7 @@ import org.springframework.stereotype.Service;
 import com.erp.inventariapp.DTOs.UserDTO;
 import com.erp.inventariapp.Entities.Person;
 import com.erp.inventariapp.Entities.UserApp;
+import com.erp.inventariapp.Exceptions.ResourceNotFoundException;
 import com.erp.inventariapp.Repositories.UserRepository;
 import com.erp.inventariapp.Repositories.PersonRepository;
 import com.erp.inventariapp.ServicesInterfaces.IUserService;
@@ -23,6 +24,9 @@ public class UserService implements IUserService {
    
     @Autowired
     PersonRepository personrepository;
+
+    @Autowired
+    PersonService personservice;
 
     @Override
     public List<UserDTO> findAll() {
@@ -52,12 +56,14 @@ public class UserService implements IUserService {
 
     @Override
     public void delete(Long iduser) {
-        //if (!userrepository.existsById(iduser))
-        //    throw ResourceNotFoundException("Categoría no encontrada");
+        if (!userrepository.existsById(iduser))
+            throw new ResourceNotFoundException("Usuario no encontrado");
         userrepository.deleteById(iduser);
     }
 
     private UserDTO saveOrUpdate(UserApp u, UserDTO dto) {
+        //Person person = modelMapper.map(dto.getPerson(), Person.class);
+
         u.setUsername(dto.getUsername());
 
         //u.setPassword(dto.getPassword());
@@ -69,14 +75,29 @@ public class UserService implements IUserService {
         
         u.setRoleUser(dto.getRoleUser());
         u.setState(dto.getState());
-        System.out.println("****--IDPERSON = "+dto.getIdperson()+" --****");
-        Optional<Person> optionalperson = personrepository.findById(dto.getIdperson());
+        System.out.println("****--IDPERSON = "+dto.getPerson().getIdperson()+" --****");
+        Optional<Person> optionalperson = personrepository.findById(dto.getPerson().getIdperson());
         if(optionalperson.isPresent()){
             u.setPerson(optionalperson.get());
         }else{
             System.out.println("****-- NOT FOUND PERSON --****");
+            // Crear nueva persona con los datos del DTO
+            Person person = Person.builder()
+            .typeId(dto.getPerson().getTypeId())
+            .identification(dto.getPerson().getIdentification())
+            .name(dto.getPerson().getName())
+            .adress(dto.getPerson().getAdress())
+            .email(dto.getPerson().getEmail())
+            .phone(dto.getPerson().getPhone())
+            .birthdate(dto.getPerson().getBirthdate())
+            .genre(dto.getPerson().getGenre())
+            .build();
+            
+            u.setPerson(person);             
         }
+        
         System.out.println("****-- END SAVEORCREATE --****");
+        
         return convertToDTO(userrepository.save(u));
     }
 
@@ -87,10 +108,13 @@ public class UserService implements IUserService {
         dto.setPassword(u.getPassword());
         dto.setRoleUser(u.getRoleUser());
         dto.setState(u.getState());
-        System.out.println("****-- BEFORE EVALUATE PERSON IN CONVERTODTO --****");
-        if(u.getPerson()!=null)
-            dto.setIdperson(u.getPerson().getIdperson());
-        System.out.println("****-- AFTER EVALUATE PERSON IN CONVERTODTO --****");
-            return dto;
+        
+        if(u.getPerson()!=null){
+            dto.setPerson(personservice.findById(u.getPerson().getIdperson()));
+        }else{
+            dto.setPerson(null);
+        }
+
+        return dto;
     }    
 }
